Renamed some files/directories to replace bst_list_view with just bst.
Also modified a comment in a test.

Files checked in: DataRepo/static/js/bst_list_view/base.js DataRepo/static/js/bst_list_view/filterer.js DataRepo/static/js/bst_list_view/sorter.js DataRepo/static/js/bst_list_view/value.js DataRepo/templates/models/bst_list_view/base.html DataRepo/templates/models/bst_list_view/td.html DataRepo/templates/models/bst_list_view/th.html DataRepo/templates/models/bst_list_view/value.html DataRepo/templatetags/bst_list_view_tags.py DataRepo/tests/static/js/bst_list_view/test_filterer.js DataRepo/tests/static/js/bst_list_view/test_sorter.js DataRepo/tests/static/js/bst_list_view/test_value.js DataRepo/tests/static/js/tests.html DataRepo/tests/templatetags/test_bst_list_view_tags.py DataRepo/tests/views/models/bst/test_list_view.py DataRepo/views/models/bst/column/filterer/base.py DataRepo/views/models/bst/column/sorter/base.py DataRepo/static/js/bst/base.js DataRepo/static/js/bst/filterer.js DataRepo/static/js/bst/sorter.js DataRepo/static/js/bst/value.js DataRepo/templates/models/bst/base.html DataRepo/templates/models/bst/td.html DataRepo/templates/models/bst/th.html DataRepo/templates/models/bst/value.html DataRepo/templatetags/bst.py DataRepo/tests/static/js/bst/test_filterer.js DataRepo/tests/static/js/bst/test_sorter.js DataRepo/tests/static/js/bst/test_value.js DataRepo/tests/templatetags/test_bst.py
