+++ DataRepo/tests/static/js/tests.html
@@ -54,15 +54,15 @@
 
         <!-- Value Tests -->
         <script src="/DataRepo/static/js/bst/value.js"></script>
-        <script src="bst_list_view/test_value.js"></script>
+        <script src="bst/test_value.js"></script>
 
         <!-- Sorter Tests -->
         <script src="/DataRepo/static/js/bst/sorter.js"></script>
-        <script src="bst_list_view/test_sorter.js"></script>
+        <script src="bst/test_sorter.js"></script>
 
         <!-- Filterer Tests -->
         <script src="/DataRepo/static/js/bst/filterer.js"></script>
-        <script src="bst_list_view/test_filterer.js"></script>
+        <script src="bst/test_filterer.js"></script>
 
     </body>
 </html>
