+++ DataRepo/tests/templatetags/test_bst.py
@@ -1,7 +1,7 @@
 from django.db.models import CharField
 from django.urls import reverse
 
-from DataRepo.templatetags.bst_list_view_tags import (
+from DataRepo.templatetags.bst import (
     get_absolute_url,
     get_attr,
     has_attr,
